% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood.R
\docType{data}
\name{blood}
\alias{blood}
\title{Blood Loss Dataset}
\format{
A data frame with 1309 and 35 variables:
\describe{
\item{Hb.prae}{prepartum hemoglobin}
\item{MBL}{measured blood loss}
\item{EC}{erythrocyte concentrate}
\item{HOSP.7d}{length of hospitalisation after delivery}
\item{COLLOIDS}{number of colloids}
\item{SPONTANEOUS}{spontaneous delivery}
\item{VE}{vacuum delivery}
\item{SECTIO.prim}{elective cesarean delivery}
\item{SECTIO.sek}{unplanned cesarean delivery}
\item{SECTIO.not}{emergency cesarean delivery}
\item{GA}{gestational age}
\item{AGE}{maternal age}
\item{MULTIPAR}{multiparity}
\item{BMI}{body mass index}
\item{DAUER.ap}{duration of second stage labor}
\item{TWIN}{multiple fetus pregnancy}
\item{IOL}{induction of labor}
\item{IOL.48h}{induction of labor $>$ 48 hours}
\item{AIS}{chorioamnionitis}
\item{NW}{neonatal weight}
\item{RUPTUR}{uterine rupture}
\item{ATONY}{uterine atony}
\item{PLAC.RET}{retained placenta}
\item{PLAC.REST}{retained placental material}
\item{PLAC.MAP}{morbidly adherent placenta}
\item{PLAC.PRAE}{placenta previa}
\item{RISS}{bleeding from laceration}
\item{PLAC.LSG}{placental abruption}
\item{F2.prae}{prepartum F. II}
\item{F1.prae}{prepartum F. I}
\item{F13.Akt.prae}{prepartum F. XIII}
\item{SUBST.F13}{substitution F. XIII}
\item{SUBST.F1}{substitution F. I}
\item{F13.Ag.prae}{prepartum F. XIII (Ag)}
\item{DD.prae}{prepartum D-dimer}
}
}
\usage{
blood
}
\description{
The blood loss dataset is based on a study by the University Hospital Zurich from October 2015 to
November 2016.
The outcome is defined as the measured postpartum blood loss \code{MBL} in mL.
Details are given in Haslinger et al. (2020).
}
\references{
Dandl S, Torsten H, Heidi S et al. (2021). Divide or Unite? Theoretical and Empirical Insights into
Two Strategies to Random Forest-type Heterogeneous Treatment Effect Estimation.

Haslinger C, Korte W, Hothorn T, Brun R, Greenberg C, Zimmermann R (2020). “The impact
of prepartum factor XIII activity on postpartum blood loss.” Journal of Thrombosis and
Haemostasis, 18. doi:10.1111/jth.14795.
}
\keyword{datasets}
