% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{dandl}
\alias{dandl}
\title{Study Setting and Results of Dandl et al. (2022)}
\format{
A data frame with 9600 rows and 23 variables:
\describe{
\item{setup}{Identifier for data generating process defined by p, m, t, sd, pi, model, xmodel.}
\item{repl}{Replication index (1 to 100).}
\item{p, m, t}{Propensity score, prognostic effect and treatment effect
function character names. Input for \code{dgp}.}
\item{sd, ol, model, xmodel}{Standard deviation of
normal distributed outcome, fraction of treatment effect added to prognostic
effect and names of used model for outcome and covariates.
Input for \code{dgp}.}
\item{nsim, dim, nsimtest, seed}{Number of observations, number of
covariates, number of observations for test set and seed. Input for
\code{simulate}.}
\item{mob, mobhonest, hybrid, hybridhonest, equalized, equalizedhonest,
cf, cfhonest, mobcf, mobcfhonest}{Mean squared error of different
methods for estimating treatment effects. See Dandl et al. (2021) for details.}
}
}
\description{
Dataset containing the study setting and performance results of 8 different
tree-based methods for estimating treatment effects based on Dandl et al. (2022).
Mean squared error of estimated to true treatment effect measured the performance.
Study settings were based on Nie and Wager (2020).
}
\references{
Dandl S, Haslinger C, Hothorn T, Seibold H, Sverdrup E, Wager S, Zeileis A (2022). What Makes
Forest-based Heterogeneous Treatment Effect Estimators Work?.

Nie X, Wager S (2020). “Quasi-Oracle Estimation of Heterogeneous Treatment Effects.” Biometrika. ISSN 0006-3444. doi:10.1093/biomet/asaa076. Asaa076, https://academic.oup.com/biomet/advance-article-pdf/doi/10.1093/biomet/asaa076/33788449/asaa076.pdf, URL https://doi.org/10.1093/biomet/asaa076.
}
