% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{dandl2}
\alias{dandl2}
\title{Study Setting and Results of Dandl et al. (2022b)}
\format{
A data frame with 41600 rows and 24 variables:
\describe{
\item{setup}{Identifier for data generating process defined by p, m, t, sd, pi, model, xmodel.}
\item{repl}{Replication index (1 to 100).}
\item{p, m, t}{Propensity score, prognostic effect and treatment effect
function character names. Input for \code{dgp}.}
\item{sd, ol, model, xmodel, rmvar}{Standard deviation of
normal distributed outcome, fraction of treatment effect added to prognostic
effect, names of used model for outcome and covariates and variable name
to be removed after simulating data.
Input for \code{dgp}.}
\item{nsim, dim, nsimtest, seed}{Number of observations, number of
covariates, number of observations for test set and seed. Input for
\code{simulate}.}
\item{mob, hybrid, equalized, equalizedgao, hybridgao, mobcox,
hybridcox, equalizedcox, equalizedgaocox, hybridgaocox}{Mean squared error of different
methods for estimating treatment effects. See Dandl et al. (2022b) for details.}
}
}
\description{
A dataset containing the study setting and performance results of five different
tree-based methods for estimating treatment effects based on Dandl et al. (2022b).
Mean squared error of estimated to true treatment effect measured the performance.
Study settings were based on Athey and Wager (2018) and  Nie and Wager (2020).
}
\references{
Dandl S, Bender A, Hothorn T (2022b). Heterogeneous
Treatment Effect Estimation for Observational Data using Model-based Forests.

Dandl S, Haslinger C, Hothorn T, Seibold H, Sverdrup E, Wager S, Zeileis A (2022a). What Makes
Forest-based Heterogeneous Treatment Effect Estimators Work?.

Wager S and Athey S. Estimation and inference of heterogeneous treatment
effects using random forests. Journal of the American Statistical Association
2018. 113(523): 1228–1242.

Athey S, Tibshirani J and Wager S. Generalized random forests. The Annals of
Statistics 2019. 47(2): 1148–1178.

Nie X and Wager S. Quasi-oracle estimation of heterogeneous treatment effects.
Biometrika 2021. 108: 299–319.
}
