% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mF}
\alias{mF}
\alias{mF_x1}
\alias{mF_x3}
\alias{mF_sin_x1_x5}
\alias{mF_max_x1_x5}
\alias{mF_log_x1_x3}
\alias{mF_max2_x1_x5}
\title{Prognostic effect functions}
\usage{
mF_x1(x)

mF_x3(x)

mF_sin_x1_x5(x)

mF_max_x1_x5(x)

mF_log_x1_x3(x)

mF_max2_x1_x5(x)
}
\arguments{
\item{x}{(data.frame)}
}
\value{
vector of corresponding prognostic effects
}
\description{
Prognostic effect functions are based on Wager and Athey (2018) and Nie and Wager (2020).
}
\details{
\code{mF_x1} corresponds to \deqn{2x_1 - 1}{2 x_1 - 1.}
Equivalently, \code{mF_x3} corresponds to \deqn{2x_3 - 1}{2 x3 - 1.}
\code{mF_sin_x1_x5} is equal to \deqn{\sin{\pi x_1 x_2} + 2(x_3 - 0.5)^2 + x_4 + 0.5 x_5}{sin(\pi x1 x2) + 2 (x3 - 0.5)^2 + x4 + 0.5 x5,}
\code{mF_max_x1_x5} to \deqn{\max{x_1 + x_2, x_3, 0} + \max{x_4 + x_5, 0}}{max(x1 + x2 + x3, 0) + max(x4 + x5, 0),}
\code{mF_log_x1_x3} to \deqn{2 \log{1 + exp(x_1 + x_2 + x_3)}}{2 log(1 + x1 + x2 + x3),}
and \code{mF_max2_x1_x5} to \deqn{(\max{x_1 + x_2 + x_3, 0} + \max{x_4 + x_5, 0})/2}{(max(x1 + x2 + x3, 0) + max(x4 + x5, 0))/2.}
}
\references{
Wager S, and Athey S (2018). "Estimation and Inference of Heterogeneous Treatment Effects using Random Forests". Journal of the American Statistical Association, 113(523).

Nie X, Wager S (2020). “Quasi-Oracle Estimation of Heterogeneous Treatment Effects.” Biometrika. ISSN 0006-3444. doi:10.1093/biomet/asaa076. Asaa076, https://academic.oup.com/biomet/advance-article-pdf/doi/10.1093/biomet/asaa076/33788449/asaa076.pdf, URL https://doi.org/10.1093/biomet/asaa076.
}
