% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pF}
\alias{pF}
\alias{pF_x1}
\alias{pF_x3}
\alias{pF_x4}
\alias{pF_sin_x3}
\alias{pF_eta_x1_x2}
\alias{pF_x2_x3}
\alias{pF_exp_x1_x2}
\title{Treatment propensity functions}
\usage{
pF_x1(x)

pF_x3(x)

pF_x4(x)

pF_sin_x3(x)

pF_eta_x1_x2(x, eta = 0.1)

pF_x2_x3(x)

pF_exp_x1_x2(x)
}
\arguments{
\item{x}{(data.frame)}

\item{eta}{(numeric(1))  Trimming parameter for \code{pF_eta_x1_x2}.}
}
\value{
vector of corresponding propensities
}
\description{
Treatment propensity functions are based on Wager and Athey (2018) and Nie and Wager (2020).
}
\details{
\code{pF_x1} corresponds to \deqn{\frac{1}{4}(1 + \beta_{2, 4}(x_1))}{1/4 (1 + \beta(x1|2, 4)).}
Equivalently, \code{pF_x3} corresponds to \deqn{\frac{1}{4}(1 + \beta_{2, 4}(x_3))}{1/4 (1 + \beta(x3|2, 4))}
and \code{pF_x4} to \deqn{\frac{1}{4}(1 + \beta_{2, 4}(x_4))}{1/4 (1 + \beta(x4|2, 4)).}
\code{pF_sin_x3} is equal to \deqn{\sin{2 * pi * x_3}/4 + 0.5}{sin(2 pi x3)/4 + 0.5,}
\code{pF_eta_x1_x2} to \deqn{\max{\eta, \min{\sin{\pi x_1 x_2}, 1 - \eta}}}{max(\eta, min(sin(\pi x1 x2), 1 - \eta)),}
\code{pF_x2_x3} to \deqn{1/(1 + \exp{x_2 + x_3})}{1/(1 + exp(x2 + x3))}
and \code{pF_exp_x1_x2} to \deqn{1/(1 + \exp{- x_1} + \exp{- x_2})}{1/(1 + exp(-x1) + exp(-x2)).}
}
\references{
Wager S, and Athey S (2018). "Estimation and Inference of Heterogeneous Treatment Effects using Random Forests". Journal of the American Statistical Association, 113(523).

Nie X, Wager S (2020). “Quasi-Oracle Estimation of Heterogeneous Treatment Effects.” Biometrika. ISSN 0006-3444. doi:10.1093/biomet/asaa076. Asaa076, https://academic.oup.com/biomet/advance-article-pdf/doi/10.1093/biomet/asaa076/33788449/asaa076.pdf, URL https://doi.org/10.1093/biomet/asaa076.
}
