% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{predict.dgp}
\alias{predict.dgp}
\title{Predict ground truth effects for new data}
\usage{
\method{predict}{dgp}(object, newdata, ...)
}
\arguments{
\item{object}{(list/gpd) Manual for data generation.}

\item{newdata}{(data.frame) New data to predict on.}
}
\value{
data.frame with true values of
pfct (pi(x)), mfct (mu(x)), tfct (tau(x)) and sdfct (sd for normal model).
}
\description{
Predict ground truth effects for new data
}
