% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{predict}
\alias{predict}
\alias{predict.dgp}
\alias{predict.simdgp}
\title{Predict ground truth effects for new data}
\usage{
\method{predict}{dgp}(object, newdata)

\method{predict}{simdgp}(object, newdata)
}
\arguments{
\item{object}{(dgp/simdgp) Manual for data generation (\code{dgp}) or generated dataset (\code{simdgp}).}

\item{newdata}{(data.frame) New data to predict on.}
}
\value{
data.frame with true values of
pfct (pi(x)), mfct (mu(x)), tfct (tau(x)) and sdfct (sd for normal model).
}
\description{
Predict ground truth effects for new data
}
\examples{
# Wager and Athey (2018) - first experiment
dgp1 <- dgp(p = pF_x1, m = mF_x1, t = 0, model = "normal", xmodel = "unif")
sim1 <- simulate(dgp1, nsim = 500L, d = 2, nsimtest = 1000L)
newdata <- attr(sim1, "testxdf")
pred1 <- predict(dgp1, newdata)
head(pred1)
pred2 <- predict(sim1, newdata)
all.equal(pred1, pred2)
}
