% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{predict.simdgp}
\alias{predict.simdgp}
\title{Predict ground truth effects for new data}
\usage{
\method{predict}{simdgp}(object, newdata, ...)
}
\arguments{
\item{object}{(list/gpd) Manual for data generation.}

\item{newdata}{(data.frame) New data to predict on.}
}
\value{
data.frame of class simdpg with columns: x, y and trt/w.
}
\description{
Predict ground truth effects for new data
}
