% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prognosticfunc}
\alias{prognosticfunc}
\alias{mF_x1}
\alias{mF_x3}
\alias{mF_sin_x1_x5}
\alias{mF_max_x1_x5}
\alias{mF_log_x1_x3}
\alias{mF_max2_x1_x5}
\title{Prognostic effect functions}
\usage{
mF_x1(x)

mF_x3(x)

mF_sin_x1_x5(x)

mF_max_x1_x5(x)

mF_log_x1_x3(x)

mF_max2_x1_x5(x)
}
\arguments{
\item{x}{(data.frame)}
}
\value{
vector of corresponding prognostic effects
}
\description{
Prognostic effect functions
}
