% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{propensityfunc}
\alias{propensityfunc}
\alias{pF_x1}
\alias{pF_x3}
\alias{pF_x4}
\alias{pF_sin_x3}
\alias{pF_eta_x1_x2}
\alias{pF_x2_x3}
\alias{pF_exp_x1_x2}
\title{Treatment propensity functions}
\usage{
pF_x1(x)

pF_x3(x)

pF_x4(x)

pF_sin_x3(x)

pF_eta_x1_x2(x, eta = 0.1)

pF_x2_x3(x)

pF_exp_x1_x2(x)
}
\arguments{
\item{x}{(data.frame)}
}
\value{
vector of corresponding propensities
}
\description{
Treatment propensity functions
}
