% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tF}
\alias{tF}
\alias{tF_exp_x1_x2}
\alias{tF_div_x1_x2}
\alias{tF_log_x1_x2}
\alias{tF_max_x1_x5}
\title{Treatment effect functions}
\usage{
tF_exp_x1_x2(x)

tF_div_x1_x2(x)

tF_log_x1_x2(x)

tF_max_x1_x5(x)
}
\arguments{
\item{x}{(data.frame)}
}
\value{
vector of corresponding treatment effects
}
\description{
Treatment effect functions are based on Wager and Athey (2018) and Nie and Wager (2020).
}
\details{
\code{tF_exp_x1_x2} corresponds to \deqn{\prod_{j = 1}^2 1 + \frac{1}{1 + \exp{(-20(x_j - \frac{1}{3}))}}}{\prod_{j = 1}^2 (1 +  1/(1 + exp(-20 (xj - (1/3))))).}
\code{tF_div_x1_x2} corresponds to \deqn{\frac{x_1 + x_2}{2}}{(x1 + x2)/2,}
\code{tF_log_x1_x2} to \deqn{x_1 + \log{1 + \exp{x_2}}}{x1 + log(1 + x2)}
and \code{tF_max_x1_x5} to \deqn{\max{x_1 + x_2 + x_3, 0} - \max{x_4 + x_5, 0}}{max(x1 + x2 + x3, 0) - max(x4 + x5, 0).}
}
\references{
Wager S, and Athey S (2018). "Estimation and Inference of Heterogeneous Treatment Effects using Random Forests". Journal of the American Statistical Association, 113(523).

Nie X, Wager S (2020). “Quasi-Oracle Estimation of Heterogeneous Treatment Ef- fects.” Biometrika. ISSN 0006-3444. doi:10.1093/biomet/asaa076. Asaa076, https://academic.oup.com/biomet/advance-article-pdf/doi/10.1093/biomet/asaa076/33788449/asaa076.pdf, URL https://doi.org/10.1093/biomet/asaa076.
}
