% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{treatmentfunc}
\alias{treatmentfunc}
\alias{tF_exp_x1_x2}
\alias{tF_div_x1_x2}
\alias{tF_log_x1_x2}
\alias{tF_max_x1_x5}
\title{Treatment effect functions}
\usage{
tF_exp_x1_x2(x)

tF_div_x1_x2(x)

tF_log_x1_x2(x)

tF_max_x1_x5(x)

tF_max_x1_x5(x)
}
\arguments{
\item{x}{(data.frame)}
}
\value{
vector of corresponding treatment effects
}
\description{
Treatment effect functions
}
\details{
\code{tF_exp_x1_x2} corresponds to \eqn{\prod_{j = 1}^2 1 + \frac{1}{1 + \exp{(-20(x_j - \frac{1}{3}))}}}.
\code{tF_div_x1_x2} corresponds to \eqn{\frac{x_1 + x_2}{2}}.
\code{tF_log_x1_x2} to \eqn{x_1 + log(1 + \exp{x_2})}
and \code{tF_max_x1_x5} to \eqn{\max{x_1 + x_2 + x_3, 0} - \max{x_4 + x_5, 0}}.
}
