% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{dgp}
\alias{dgp}
\title{Create manual for data generation}
\usage{
dgp(
  p = 0.5,
  m = 0,
  t = 0,
  sd = 1,
  pi = 0.5,
  model = c("normal", "weibull", "binomial", "polr"),
  xmodel = c("normal", "unif")
)
}
\arguments{
\item{p}{(numeric(1), function) Propensities P(trt = 1|x) (pi(x)),
default = 0.5 means that there is no confounding.}

\item{m}{(numeric(1), function) Prognostic effect (mu(x)), default 0 means that
there is no prognostic effect.}

\item{t}{(numeric(1), function) Predictive treatment effect (tau(x)),
default 0 means that there is no treatment effect.}

\item{sd}{(numeric(1)) Standard deviation of normal distribution, only
has an effect if model = "normal".}

\item{pi}{(numeric(1)) how much of predictive effect is added to prognostic effect.
If pi = 0, conditional mean does not depend on treatment effect.}

\item{model}{("normal"|"weibull") Name of used model.}
}
\value{
list of class gpd with entries:
pfct (pi(x)), mfct (mu(x)), tfct (tau(x)), sdfct (sd for normal model),
model (model used)).
}
\description{
Create manual for data generation
}
