% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{simulate.dgp}
\alias{simulate.dgp}
\title{Simulate from a given dgp object}
\usage{
\method{simulate}{dgp}(object, nsim = 1, seed = NULL, dim = 4, nsimtest = NULL, ...)
}
\arguments{
\item{object}{(list/dgp) Manual for data generation.}

\item{nsim}{(numeric(1)) Number of observations (n), default 1.}

\item{seed}{(numeric(1)) Seed for data generation, default NULL.}

\item{dim}{(numeric(1)) Dimension, number of predictors (p), default 4.}

\item{nsimtest}{(numeric(1)) Number of observations (n) for test dataset.
Default NULL means that no test dataset is generated.}

\item{...}{Additional optional arguments.}
}
\value{
data.frame of class simdpg with columns: x, y and trt and attributes
\itemize{
\item \code{truth}: the  \code{object}
\item \code{testxdf}: the test data with \code{nsimtest} rows and \code{dim} columns.
Test datasets are not affected from specification of \code{rmvar} in \code{object} (see \link{dgp}).
\item \code{runseed}: an additional seed (e.g., to run a method)
}
}
\description{
Simulate data from a given \code{dgp} object.
}
\examples{

# Wager and Athey (2018) - first experiment
dgp1 <- dgp(p = pF_x1, m = mF_x1, t = 0, model = "normal", xmodel = "unif")
sim1 <- simulate(dgp1, nsim = 500L, d = 2L, nsimtest = 1000L) # in paper d = {2, 5, 10, 15, 20, 30}
head(sim1)

# Nie and Wager (2020) - Setup A
dgpA <- dgp(p = pF_eta_x1_x2, m = mF_sin_x1_x5, t = tF_div_x1_x2,
 model = "normal", xmodel = "unif")
simA <- simulate(dgpA, nsim = 500L, d = 6L, nsimtest = 500L)
head(simA)

# Get test data
testdf <- attr(simA, "testxdf")
head(testdf)

}
\seealso{
\code{\link{dgp}}
}
